dcr/scenarios/agent-bvt-module/requirements.txt
pytest
junit-xml
distro
azure-identity
azure-keyvault-keys
azure-mgmt-compute>=22.1.0
azure-mgmt-keyvault>=7.0.0
azure-mgmt-network>=16.0.0
azure-mgmt-resource>=15.0.0
python-dotenv
pudb